fix: support :deps-file, :dry-run in test-util/neil
This test helper now supports the args used by `dep_upgrade_test`, which
expected to set the :deps-file and a :dry-run flag.

It may be preferred to leave this as is, and instead build up the neil
command as a string (rather than support neil's args in this test helper).