upgrade alias extra-deps (#127)
* feat: support updating aliases in dep upgrade

Includes :aliases <alias> :extra-deps in the deps upgraded by `neil dep
upgrade`. Supports specifying a :lib, which can be in the normal :deps
or an aliases's :extra-deps map.

* feat: support filtering upgrades by alias

`dep upgrade` now respects a passed --alias, which will limit the scope
of your upgrade to any alias's extra-deps.

* docs: upgrade help text noting alias behavior.

* feat: support --no-aliases when upgrading

* docs: update changelog

* refactor: filter to specified deps in one place

opts->specified-deps seems to be better seperation of concerns, and is
perhaps more reusable/extenable now.