fix: handle missing versions in dep add
My previous git/tag PR adjusted the dep add algorithm, but did not
handle missing versions. This introduced a deps.edn-clearing bug when a
lib was added that did not exist.

This is a quick fix that prevents updating the deps.end in `add-dep`
when no version is found for the passed lib.

We may also want to help the UX here with a println or some other
validation hint.