fix: print and exit when github rate limited
Handles github rate-limited responses in curl-get-json. We now print a
message and exit. Includes some docs calling out the env vars for
raising the limit.

We may prefer to nil-pun rather than exit. For now I decided to follow
the way we exit in the `license` code, which similarly binds *err* to *out*,
prints a message, and exits with an error code.