Fix #136: Handle Github rate limit reached (#137)
* fix: print and exit when github rate limited

Handles github rate-limited responses in curl-get-json. We now print a
message and exit. Includes some docs calling out the env vars for
raising the limit.

We may prefer to nil-pun rather than exit. For now I decided to follow
the way we exit in the `license` code, which similarly binds *err* to *out*,
prints a message, and exits with an error code.

* update changelog

* feat: move to friendly env vars

The existing env vars will still work for now, but that support may be
dropped in the future.

* changelog: include links to resolved issue