pick up the package name when in deps file (#195)
* pick up the package name when in deps file

when called the command, it picks up the package name automatically if
the buffer's filename is deps.edn or project.clj

* add a check

urgggh... I hate elisp sometimes. symbol-name returns "nil" if you pass
nil to it and throws an error if you give it something like an empty
string. And there's no cond-> macro in elisp.