Store preconfigured aliases as data
In https://github.com/babashka/neil/pull/216, we changed Neil's
internal `add-alias` functionality to work on aliases as Clojure data
instead of strings.

This PR follows up with cleanup.

After #216 was merged, there is no longer any reason to store aliases as
strings internally and convert to data. In this commit, we:

- Store aliases as data rather than strings (eg
`babashka.neil/kaocha-alias`)

- Change code that assumed aliases were strings to assume aliases are
  data.