Store provided aliases as data instead of strings (#221)
* alias-map can be str or data

* Store preconfigured aliases as data

In https://github.com/babashka/neil/pull/216, we changed Neil's
internal `add-alias` functionality to work on aliases as Clojure data
instead of strings.

This PR follows up with cleanup.

After #216 was merged, there is no longer any reason to store aliases as
strings internally and convert to data. In this commit, we:

- Store aliases as data rather than strings (eg
`babashka.neil/kaocha-alias`)

- Change code that assumed aliases were strings to assume aliases are
  data.

* Refactor: kaocha tests have kaocha in test name

* Refactor: don't pass in unused argument

* Delete unused argument

`:s` is the build alias as a string. The build alias as string is no
longer in use.

* Refactor: -latest suffix for alias fns

Motivation: these functions look up the latest version of the
coordinates for certain aliases. I think the previous name was more
confusing.

* Delete unused local variable